Move command_args_mixin to _grpc_services.
