+++ tests/test_datamodel_service.py
@@ -33,12 +33,12 @@ from ansys.api.fluent.v0 import datamodel_se_pb2
 from ansys.api.fluent.v0.variant_pb2 import Variant
 import ansys.fluent.core as pyfluent
 from ansys.fluent.core import examples
+from ansys.fluent.core._grpc_services._command_arguments_mixin import (
+    CommandArgumentsCleanupMixin,
+)
 from ansys.fluent.core._grpc_services.object_model_service_v0 import (
     _convert_value_to_variant,
     _convert_variant_to_value,
-)
-from ansys.fluent.core.services._command_arguments_mixin import (
-    CommandArgumentsCleanupMixin,
 )
 from ansys.fluent.core.services.object_model import (
     PyArguments,
